+++ apps/admin-api/src/override/override.service.ts
@@ -1,0 +1,55 @@
+import { Injectable } from '@nestjs/common'
+import { InjectModel } from '@nestjs/mongoose'
+
+import { StudyProgram } from '@thinc-org/chula-courses'
+import { Model } from 'mongoose'
+
+import { OverrideInput } from '../graphql'
+import { Override, OverrideDocument } from '../schemas/override.schema'
+
+@Injectable()
+export class OverrideService {
+  constructor(
+    @InjectModel('override')
+    private overrideModel: Model<OverrideDocument>
+  ) {}
+
+  async getOverrides(): Promise<Override[]> {
+    const result = await this.overrideModel.find().lean()
+    return result
+  }
+
+  async createOrUpdateOverride(override: OverrideInput): Promise<Override> {
+    const result = await this.overrideModel
+      .findOneAndUpdate(
+        {
+          courseNo: override.courseNo,
+          studyProgram: override.studyProgram,
+          academicYear: override.academicYear,
+          semester: override.semester,
+        },
+        override,
+        {
+          new: true,
+          upsert: true,
+        }
+      )
+      .lean()
+    return result
+  }
+
+  async deleteOverride(
+    courseNo: string,
+    studyProgram: StudyProgram,
+    academicYear: string,
+    semester: string
+  ): Promise<Override> {
+    const result = await this.overrideModel.findOneAndDelete({
+      courseNo,
+      studyProgram,
+      academicYear,
+      semester,
+    })
+    return result
+  }
+}