+++ apps/admin-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # admin-api
 
+## 0.1.0-beta.0
+
+### Minor Changes
+
+- e1b5095b: upgrade to node 18
+- 954e6c34: Add review and override to admin-api
+
 ## 0.0.4
 
 ### Patch Changes